Use alternative_sync_chidren in tap_core
