+++ tap_github/tests/fixtures.py
@@ -1,3 +1,5 @@
+import os
+import logging
 import datetime
 
 import pytest
@@ -25,7 +27,7 @@ def repo_list_config(request):
     """
     marker = request.node.get_closest_marker("repo_list")
     if marker is None:
-        repo_list = ["octocat/hello-world", "mapswipe/mapswipe"]
+        repo_list = ["MeltanoLabs/tap-github", "mapswipe/mapswipe"]
     else:
         repo_list = marker.args[0]
 
@@ -93,3 +95,28 @@ def organization_list_config(request):
         "organizations": organization_list,
         "rate_limit_buffer": 100,
     }
+
+
+def alternative_sync_chidren(self, child_context: dict) -> None:
+    """
+    Override for Stream._sync_children.
+    Enabling us to use an ORG_LEVEL_TOKEN for the collaborators sream.
+    """
+    for child_stream in self.child_streams:
+        # Use org:write access level credentials for collaborators stream
+        if child_stream.name in ["collaborators"]:
+            ORG_LEVEL_TOKEN = os.environ.get("ORG_LEVEL_TOKEN")
+            if not ORG_LEVEL_TOKEN:
+                logging.warning(
+                    'No "ORG_LEVEL_TOKEN" found. Skipping collaborators stream sync.'
+                )
+                continue
+            SAVED_GTHUB_TOKEN = os.environ.get("GITHUB_TOKEN")
+            os.environ["GITHUB_TOKEN"] = ORG_LEVEL_TOKEN
+            child_stream.sync(context=child_context)
+            os.environ["GITHUB_TOKEN"] = SAVED_GTHUB_TOKEN or ""
+            continue
+
+        # default behavior:
+        if child_stream.selected or child_stream.has_selected_descendents:
+            child_stream.sync(context=child_context)
